+++ CHANGES.rst
@@ -35,6 +35,7 @@ Development
 - Improve log messages
 - Bump version to 0.4.0dev
 - Improve purging of sensible configuration settings
+- SystemBatteryLevel: Obtain voltage divider parameters from settings
 
 
 2019-06-07 0.3.0